requirements.txt
musicbrainzngs>=0.7.1
lyricsgenius>=3.0.1
python-slugify>=8.0
pyyaml>=6.0
